chore(): review comments and polishing
