+++ package.json
@@ -28,7 +28,9 @@
     "lucide-react": "^0.438.0"
   },
   "devDependencies": {
+    "@babel/preset-env": "^7.26.7",
     "@babel/preset-react": "^7.24.1",
+    "@babel/preset-typescript": "^7.26.0",
     "@rollup/plugin-node-resolve": "^15.2.3",
     "@rollup/plugin-typescript": "^11.1.6",
     "@rsbuild/core": "^1.0.1",
