+++ src/Anar.css
@@ -7,7 +7,7 @@
 
 @import './Text/Label.css';
 @import './Button/Button.css';
-@import './Core/CheckBox.css';
+@import './Core/Checkbox.css';
 @import './Radio/Radio.css';
 @import './Radio/RadioGroup.css';
